+++ src/icons/organization-icon.tsx
@@ -1,0 +1,27 @@
+/*
+ * Copyright (c) 2011-2020 Genestack Limited
+ * All Rights Reserved
+ * THIS IS UNPUBLISHED PROPRIETARY SOURCE CODE OF GENESTACK LIMITED
+ * The copyright notice above does not evidence any
+ * actual or intended publication of such source code.
+ */
+
+import * as React from 'react';
+
+import {createIcon} from '../components/icon';
+
+// tslint:disable max-line-length
+
+/**
+ * Building icon
+ */
+export const OrganizationIcon = createIcon(
+    <svg viewBox="0 0 20 20">
+        <path
+            fillRule="evenodd"
+            clipRule="evenodd"
+            d="M8 3V4H7C6.44772 4 6 4.44772 6 5V16H4V17H6.5H14.5H17V16H15V5C15 4.44772 14.5523 4 14 4H13V3H8ZM11 16H7V5L14 5V16H13V12.5C13 12.2239 12.7761 12 12.5 12H11.5C11.2239 12 11 12.2239 11 12.5V16ZM8 6.5C8 6.22386 8.22386 6 8.5 6H9.5C9.77614 6 10 6.22386 10 6.5V7.5C10 7.77614 9.77614 8 9.5 8H8.5C8.22386 8 8 7.77614 8 7.5V6.5ZM11.5 6C11.2239 6 11 6.22386 11 6.5V7.5C11 7.77614 11.2239 8 11.5 8H12.5C12.7761 8 13 7.77614 13 7.5V6.5C13 6.22386 12.7761 6 12.5 6H11.5ZM11 9.5C11 9.22386 11.2239 9 11.5 9H12.5C12.7761 9 13 9.22386 13 9.5V10.5C13 10.7761 12.7761 11 12.5 11H11.5C11.2239 11 11 10.7761 11 10.5V9.5ZM8 9.5C8 9.22386 8.22386 9 8.5 9H9.5C9.77614 9 10 9.22386 10 9.5V10.5C10 10.7761 9.77614 11 9.5 11H8.5C8.22386 11 8 10.7761 8 10.5V9.5ZM8.5 12C8.22386 12 8 12.2239 8 12.5V13.5C8 13.7761 8.22386 14 8.5 14H9.5C9.77614 14 10 13.7761 10 13.5V12.5C10 12.2239 9.77614 12 9.5 12H8.5Z"
+            className="gs-svg-icon-fill"
+        />
+    </svg>
+);